+++ packages/validator/source/service-provider.ts
@@ -8,6 +8,7 @@ import { assert } from "@mainsail/utils";
 import Joi from "joi";
 
 import { BIP38, BIP39 } from "./keys/index.js";
+import { TransactionForger } from "./transaction-forger.js";
 import { ValidatorRepository } from "./validator-repository.js";
 import { Validator } from "./validator.js";
 
@@ -15,6 +16,7 @@ import { Validator } from "./validator.js";
 export class ServiceProvider extends Providers.ServiceProvider {
 	public async register(): Promise<void> {
 		this.app.bind(Identifiers.Validator.Repository).to(ValidatorRepository).inSingletonScope();
+		this.app.bind(Identifiers.Validator.TransactionForger).to(TransactionForger).inSingletonScope();
 	}
 
 	public async boot(): Promise<void> {
